# util/collate: implement utf8mb4_0900_ai_ci collation | tidb-test=pr/2192