--- a/.gitattributes
+++ b/.gitattributes
@@ -3,3 +3,6 @@
 
 # Declare files that will always have LF line endings on checkout.
 *.y text eol=lf
+
+util/collate/unicode_0*_ci.go linguist-generated=true
+util/collate/ucadata/unicode_*_data.go linguist-generated=true