--- a/cmd/explaintest/r/collation_misc_enabled.result
+++ b/cmd/explaintest/r/collation_misc_enabled.result
@@ -106,6 +106,7 @@ latin1	47	1
 utf8	83	1
 utf8	33	1
 utf8	192	1
+utf8mb4	255	1
 utf8mb4	46	1
 utf8mb4	45	1
 utf8mb4	224	1
@@ -130,6 +131,7 @@ latin1_bin	latin1	47	Yes	Yes	1
 utf8_bin	utf8	83	Yes	Yes	1
 utf8_general_ci	utf8	33		Yes	1
 utf8_unicode_ci	utf8	192		Yes	1
+utf8mb4_0900_ai_ci	utf8mb4	255		Yes	1
 utf8mb4_bin	utf8mb4	46	Yes	Yes	1
 utf8mb4_general_ci	utf8mb4	45		Yes	1
 utf8mb4_unicode_ci	utf8mb4	224		Yes	1