--- a/executor/test/seqtest/seq_executor_test.go
+++ b/executor/test/seqtest/seq_executor_test.go
@@ -1202,6 +1202,7 @@ func TestShowForNewCollations(t *testing.T) {
 		"utf8_bin utf8 83 Yes Yes 1",
 		"utf8_general_ci utf8 33  Yes 1",
 		"utf8_unicode_ci utf8 192  Yes 1",
+		"utf8mb4_0900_ai_ci utf8mb4 255  Yes 1",
 		"utf8mb4_bin utf8mb4 46 Yes Yes 1",
 		"utf8mb4_general_ci utf8mb4 45  Yes 1",
 		"utf8mb4_unicode_ci utf8mb4 224  Yes 1",