--- a/util/collate/ucadata/BUILD.bazel
+++ b/util/collate/ucadata/BUILD.bazel
@@ -1,0 +1,25 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "ucadata",
+    srcs = [
+        "data.go",
+        "unicode_0900_ai_ci_data_generated.go",
+        "unicode_ci_data_generated.go",
+    ],
+    importpath = "github.com/pingcap/tidb/util/collate/ucadata",
+    visibility = ["//visibility:public"],
+)
+
+go_test(
+    name = "ucadata_test",
+    timeout = "short",
+    srcs = [
+        "unicode_0900_ai_ci_data_test.go",
+        "unicode_ci_data_original_test.go",
+        "unicode_ci_data_test.go",
+    ],
+    embed = [":ucadata"],
+    flaky = True,
+    deps = ["@com_github_stretchr_testify//require"],
+)