… build/nogo_config.json
… expression/builtin_string_test.go
… expression/collation_test.go
… expression/test/collation/uca_test.go
… util/collate/BUILD.bazel
… util/collate/collate_bench_test.go
… util/collate/collate_test.go
… util/collate/ucadata/generator/allkeys-4.0.0.txt
… util/collate/ucadata/generator/allkeys-9.0.0.txt
… util/collate/ucadata/generator/data.go.tpl
… util/collate/ucadata/generator/magic.go
… util/collate/ucadata/unicode_0900_ai_ci_data_generated.go
… util/collate/ucadata/unicode_0900_ai_ci_data_test.go
… util/collate/ucadata/unicode_ci_data_generated.go
… util/collate/ucadata/unicode_ci_data_original_test.go
… util/collate/ucadata/unicode_ci_data_test.go
… util/collate/ucaimpl/unicode_ci.go.tpl
… util/collate/unicode_0400_ci_generated.go
… util/collate/unicode_0900_ai_ci_generated.go