### What problem does this PR solve?

Issue Number: close #37566

Problem Summary:

Add `utf8mb4_0900_ai_ci` collation to the TiDB.

This PR is 400-600 lines long, quite big. (A lot of lines in the 100k+ are generated or downloaded from unicode.org, so don't be too scared).

### What is changed and how it works?

1. Refractor the original `unicode_ci` implementation to include the code generator in the source files. (Also add tests to make sure they are still the same).
2. Generate data codes for the `utf8mb4_0900_ai_ci` collation.
3. Re-use most of the implementation from `unicode_ci`, through code generation.

This PR uses code generation in two places: one for the data (from allkeys.txt), one for the implementation (to reuse the existing `unicode_ci` collation).

The second choice is hard, a more normal way is to use `interface` or something to provide different tiny implementation based on a more general UCA framework. However, as I've tested, if the related functions are not inlined (especially when it's an interface method call, or a generic method call), the performance can degrade 20%-50%.

### TODO

- [ ] Make sure the generated table is the same with MySQL's in `uca_data.h`.

### Benchmark

It becomes faster, and the `0900AICI` is faster than `UnicodeCI` in the benchmark, but I don't know why 🤦 .

#### Before this PR

```
BenchmarkUtf8mb4UnicodeCI_CompareShort-8        220942854              319.9 ns/op
BenchmarkUtf8mb4UnicodeCI_CompareMid-8           3252583             22085 ns/op
BenchmarkUtf8mb4UnicodeCI_CompareLong-8             2590          28277470 ns/op
BenchmarkUtf8mb4UnicodeCI_KeyShort-8            362425852              207.8 ns/op
BenchmarkUtf8mb4UnicodeCI_KeyMid-8               5530659             13061 ns/op
BenchmarkUtf8mb4UnicodeCI_KeyLong-8                 4460          15526126 ns/op
```

#### After this PR

```
BenchmarkUtf8mb4UnicodeCI_CompareShort-8        232078950              302.8 ns/op
BenchmarkUtf8mb40900AICI_CompareShort-8         235785336         296.0 ns/op
BenchmarkUtf8mb4UnicodeCI_CompareMid-8           3324604             21180 ns/op
BenchmarkUtf8mb40900AICI_CompareMid-8            3447740             20757 ns/op
BenchmarkUtf8mb4UnicodeCI_CompareLong-8             2659          27342506 ns/op
BenchmarkUtf8mb40900AICI_CompareLong-8              2706          26881637 ns/op
BenchmarkUtf8mb4UnicodeCI_KeyShort-8            368781840              195.1 ns/op
BenchmarkUtf8mb40900AICI_KeyShort-8             400789008              182.1 ns/op
BenchmarkUtf8mb4UnicodeCI_KeyMid-8               5853435             12336 ns/op
BenchmarkUtf8mb40900AICI_KeyMid-8                6267391             11592 ns/op
BenchmarkUtf8mb4UnicodeCI_KeyLong-8                 4704          15383704 ns/op
BenchmarkUtf8mb40900AICI_KeyLong-8                  4426          15551595 ns/op
```

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [x] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No code

### Release note

```release-note
Support the `utf8mb4_0900_ai_ci` collation.
```