@dveeden 2023-08-03
```sql
select 'foo' collate utf8mb4_0900_ai_ci = 'Foo' collate utf8mb4_general_ci
```

results in:

```
[2023/08/03 08:19:12.178 +02:00] [WARN] [session.go:2255] ["compile SQL failed"] [conn=2097156] [error="[expression:1267]Illegal mix of collations (utf8mb4_0900_ai_ci,EXPLICIT) and (utf8mb4_general_ci,EXPLICIT) for operation '='"] [SQL="select 'foo' collate utf8mb4_0900_ai_ci = 'Foo' collate utf8mb4_general_ci"]
[2023/08/03 08:19:12.178 +02:00] [INFO] [conn.go:1082] ["command dispatched failed"] [conn=2097156] [connInfo="id:2097156, addr:127.0.0.1:60728 status:10, collation:utf8mb4_0900_ai_ci, user:root"] [command=Query] [status="inTxn:0, autocommit:1"] [sql="select 'foo' collate utf8mb4_0900_ai_ci = 'Foo' collate utf8mb4_general_ci"] [txn_mode=PESSIMISTIC] [timestamp=0] [err="[expression:1267]Illegal mix of collations (utf8mb4_0900_ai_ci,EXPLICIT) and (utf8mb4_general_ci,EXPLICIT) for operation '='\ngithub.com/pingcap/errors.AddStack\n\t/home/dvaneeden/go/pkg/mod/github.com/pingcap/errors@v0.11.5-0.20221009092201-b66cddb77c32/errors.go:174\ngithub.com/pingcap/errors.(*Error).GenWithStackByArgs\n\t/home/dvaneeden/go/pkg/mod/github.com/pingcap/errors@v0.11.5-0.20221009092201-b66cddb77c32/normalize.go:164\ngithub.com/pingcap/tidb/expression.illegalMixCollationErr\n\t/home/dvaneeden/dev/pingcap/tidb/expression/collation.go:508\ngithub.com/pingcap/tidb/expression.CheckAndDeriveCollationFromExprs\n\t/home/dvaneeden/dev/pingcap/tidb/expression/collation.go:306\ngithub.com/pingcap/tidb/expression.deriveCollation\n\t/home/dvaneeden/dev/pingcap/tidb/expression/collation.go:217\ngithub.com/pingcap/tidb/expression.newBaseBuiltinFuncWithTp\n\t/home/dvaneeden/dev/pingcap/tidb/expression/builtin.go:152\ngithub.com/pingcap/tidb/expression.(*compareFunctionClass).generateCmpSigs\n\t/home/dvaneeden/dev/pingcap/tidb/expression/builtin_compare.go:1775\ngithub.com/pingcap/tidb/expression.(*compareFunctionClass).getFunction\n\t/home/dvaneeden/dev/pingcap/tidb/expression/builtin_compare.go:1769\ngithub.com/pingcap/tidb/expression.newFunctionImpl\n\t/home/dvaneeden/dev/pingcap/tidb/expression/scalar_function.go:232\ngithub.com/pingcap/tidb/expression.NewFunction\n\t/home/dvaneeden/dev/pingcap/tidb/expression/scalar_function.go:263\ngithub.com/pingcap/tidb/planner/core.(*expressionRewriter).newFunction\n\t/home/dvaneeden/dev/pingcap/tidb/planner/core/expression_rewriter.go:1400\ngithub.com/pingcap/tidb/planner/core.(*expressionRewriter).constructBinaryOpFunction\n\t/home/dvaneeden/dev/pingcap/tidb/planner/core/expression_rewriter.go:278\ngithub.com/pingcap/tidb/planner/core.(*expressionRewriter).binaryOpToExpression\n\t/home/dvaneeden/dev/pingcap/tidb/planner/core/expression_rewriter.go:1543\ngithub.com/pingcap/tidb/planner/core.(*expressionRewriter).Leave\n\t/home/dvaneeden/dev/pingcap/tidb/planner/core/expression_rewriter.go:1259\ngithub.com/pingcap/tidb/parser/ast.(*BinaryOperationExpr).Accept\n\t/home/dvaneeden/dev/pingcap/tidb/parser/ast/expressions.go:223\ngithub.com/pingcap/tidb/planner/core.(*PlanBuilder).rewriteExprNode\n\t/home/dvaneeden/dev/pingcap/tidb/planner/core/expression_rewriter.go:201\ngithub.com/pingcap/tidb/planner/core.(*PlanBuilder).rewriteWithPreprocess\n\t/home/dvaneeden/dev/pingcap/tidb/planner/core/expression_rewriter.go:146\ngithub.com/pingcap/tidb/planner/core.(*PlanBuilder).buildProjection\n\t/home/dvaneeden/dev/pingcap/tidb/planner/core/logical_plan_builder.go:1724\ngithub.com/pingcap/tidb/planner/core.(*PlanBuilder).buildSelect\n\t/home/dvaneeden/dev/pingcap/tidb/planner/core/logical_plan_builder.go:4493\ngithub.com/pingcap/tidb/planner/core.(*PlanBuilder).Build\n\t/home/dvaneeden/dev/pingcap/tidb/planner/core/planbuilder.go:853\ngithub.com/pingcap/tidb/planner.buildLogicalPlan\n\t/home/dvaneeden/dev/pingcap/tidb/planner/optimize.go:559\ngithub.com/pingcap/tidb/planner.optimize\n\t/home/dvaneeden/dev/pingcap/tidb/planner/optimize.go:479\ngithub.com/pingcap/tidb/planner.Optimize\n\t/home/dvaneeden/dev/pingcap/tidb/planner/optimize.go:337\ngithub.com/pingcap/tidb/executor.(*Compiler).Compile\n\t/home/dvaneeden/dev/pingcap/tidb/executor/compiler.go:98\ngithub.com/pingcap/tidb/session.(*session).ExecuteStmt\n\t/home/dvaneeden/dev/pingcap/tidb/session/session.go:2225\ngithub.com/pingcap/tidb/server.(*TiDBContext).ExecuteStmt\n\t/home/dvaneeden/dev/pingcap/tidb/server/driver_tidb.go:291\ngithub.com/pingcap/tidb/server.(*clientConn).handleStmt\n\t/home/dvaneeden/dev/pingcap/tidb/server/conn.go:1994\ngithub.com/pingcap/tidb/server.(*clientConn).handleQuery\n\t/home/dvaneeden/dev/pingcap/tidb/server/conn.go:1785\ngithub.com/pingcap/tidb/server.(*clientConn).dispatch\n\t/home/dvaneeden/dev/pingcap/tidb/server/conn.go:1272\ngithub.com/pingcap/tidb/server.(*clientConn).Run\n\t/home/dvaneeden/dev/pingcap/tidb/server/conn.go:1051\ngithub.com/pingcap/tidb/server.(*Server).onConn\n\t/home/dvaneeden/dev/pingcap/tidb/server/server.go:699\nruntime.goexit\n\t/usr/lib/golang/src/runtime/asm_amd64.s:1598"]
```

I don't think logging the full stack is useful for this. However this might be out of scope for this PR.