@YangKeao 2023-08-14
> Are we sure the license of `allkeys-9.0.0.txt` is fine?

The license of unicode data file is: https://www.unicode.org/license.txt . It's seems fine.