@dveeden 2023-08-14
> > Are we sure the license of `allkeys-9.0.0.txt` is fine?
> 
> @dveeden The license of unicode data file is: https://www.unicode.org/license.txt . It's seems fine.

Yes that indeed seems to be fine. I think we should include a copy of the license in the `LICENSES/` directory.

```
$ find LICENSES/
LICENSES/
LICENSES/QL-LICENSE
LICENSES/STRUTIL-LICENSE
```