@YangKeao 2023-08-14
> select 'foo' collate utf8mb4_0900_ai_ci = 'Foo' collate utf8mb4_general_ci

@dveeden I think this error is compatible with MySQL. MySQL also gives the same error:

```
mysql> select 'foo' collate utf8mb4_0900_ai_ci = 'Foo' collate utf8mb4_general_ci;
ERROR 1267 (HY000): Illegal mix of collations (utf8mb4_0900_ai_ci,EXPLICIT) and (utf8mb4_general_ci,EXPLICIT) for operation '='
```

The document also approves this behavior:

> MySQL uses coercibility values with the following rules to resolve ambiguities: 
> - If both sides have the same coercibility, then: 
>   - If both sides are Unicode, or both sides are not Unicode, it is an error.