@YangKeao 2023-08-14
I've checked that the `weight_string` is equal between TiDB and MySQL for each characters in utf8mb4_0900_ai_ci, but the default character set of parameters has a little difference. See: https://github.com/pingcap/tidb/issues/46059

The test program:

```
package main

import (
	"database/sql"
	"fmt"

	_ "github.com/go-sql-driver/mysql"
	"golang.org/x/exp/slices"
)

func main() {
	dbMySQL, err := sql.Open("mysql", "root:123456@/test")
	if err != nil {
		panic(err)
	}
	dbTiDB, err := sql.Open("mysql", "root:@tcp(127.0.0.1:4000)/test")
	if err != nil {
		panic(err)
	}

	for i := 0; i < 183969; i++ {
		if i%1000 == 0 {
			fmt.Printf("i: %d\n", i)
		}
		rows, err := dbMySQL.Query("SELECT WEIGHT_STRING(? collate utf8mb4_0900_ai_ci)", fmt.Sprintf("%c", rune(i)))
		if err != nil {
			panic(err)
		}
		var expectedWeight []uint8

		for rows.Next() {
			err = rows.Scan(&expectedWeight)
			if err != nil {
				panic(err)
			}
			break
		}
		rows.Close()

		rows, err = dbTiDB.Query("SELECT WEIGHT_STRING(CONVERT(? USING utf8mb4) collate utf8mb4_0900_ai_ci)", fmt.Sprintf("%c", rune(i)))
		if err != nil {
			panic(err)
		}
		var weight []uint8
		for rows.Next() {
			err = rows.Scan(&weight)
			if err != nil {
				panic(err)
			}
			break
		}
		rows.Close()

		if !slices.Equal(expectedWeight, weight) {
			fmt.Printf("The weight of %X is not equal: expectedWeight: %v, weight: %v\n", i, expectedWeight, weight)
		}
	}
}
```